move LspExecuteCommand to its own file
