+++ plugin/configuration.py
@@ -8,10 +8,9 @@ from .core.configurations import (
     create_window_configs,
     get_global_client_config
 )
-from .core.registry import config_for_scope, windows, client_for_view, LspTextCommand
+from .core.registry import config_for_scope, windows
 from .core.events import global_events
 from .core.workspace import enable_in_project, disable_in_project
-from .core.protocol import Request
 
 try:
     from typing import List, Optional, Dict, Any
@@ -72,33 +71,6 @@ class LspEnableLanguageServerGloballyCommand(sublime_plugin.WindowCommand):
 
             sublime.set_timeout_async(lambda: wm.start_active_views(), 500)
             self.window.status_message("{} enabled, starting server...".format(config_name))
-
-
-class LspExecuteCommand(LspTextCommand):
-    def __init__(self, view):
-        super().__init__(view)
-
-    def run(self, lsp_command) -> None:
-        self._commands = []   # type: List[str]
-        for config in client_configs.all:
-            if config.commands:
-                self._commands.extend(config.commands)
-
-        if len(self._commands) > 0:
-            self.view.window().show_quick_panel(self._commands, self._on_done)
-
-    def _handle_response(self, response: 'Optional[Any]') -> None:
-        pass
-
-    def _on_done(self, index: int) -> None:
-        if index > -1:
-            command = self._commands[index]
-            client = client_for_view(self.view)
-            if client:
-                request = {
-                    "command": command
-                }
-                client.send_request(Request.executeCommand(request), self._handle_response)
 
 
 class LspEnableLanguageServerInProjectCommand(sublime_plugin.WindowCommand):
